Update linear_algebra.cpp
This file contains the function that performs the Biconjugate Gradient Stabilised method, which is used to find the vector x in the equation xA = b. This code works for A as a tridiagonal matrix, and takes as inputs the 3 vectors representing A, and the target vector b, and outputs the vector x.